requirements.txt
gradio
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
ddgs
python-dotenv>=1.0.0


